toolbox: pin tools used in toolbox
Only the `init` command is run with the user installed packages
from it's distribution. So it must be minimal and we must make sure it
works everywhere.

Other commands are run trough nix-shell with a pinned set of dependencies
(git, coreutils, etc...). This ensure that everyone running any command
uses the same dependencies, with the same options etc.