fly: dynamic selection of fly binary
This commit affects the behaviour of fly. Instead of calling fly
directly, a script will check the version of the Concourse server in
order to use the right fly binary.

Fly is also wrapped by a script that add `fly switch` subcommand to
change Concourse's target with a more friendly manner. It also remember
which target you are using in order to prevent adding `--target` on each
fly command.